+++ dhlab/api/dhlab_api.py
@@ -15,7 +15,13 @@ def get_metadata(urns = None):
     r = requests.post(f"{BASE_URL}/get_metadata", json = params)
     return pd.DataFrame(r.json())
 
-def get_reference(corpus = 'digavis', from_year = 1950, to_year = 1955, lang = 'nob', limit = 100000):
+def get_reference(
+    corpus = 'digavis',
+    from_year = 1950,
+    to_year = 1955,
+    lang = 'nob',
+    limit = 100000
+):
     params = locals()
     r = requests.get(BASE_URL + "/reference_corpus", params = params)
     if r.status_code == 200:
@@ -25,7 +31,8 @@ def get_reference(corpus = 'digavis', from_year = 1950, to_year = 1955, lang = '
     return pd.DataFrame(result, columns = ['word', 'freq']).set_index('word')
 
 def find_urns(docids = None, mode = 'json'):
-    """ Return a list of URNs from a list of docids as a dictionary {docid: URN} or as a pandas dataframe"""
+    """ Return a list of URNs from a list of docids as
+    a dictionary {docid: URN} or as a pandas dataframe"""
 
     params = locals()
     r = requests.post(BASE_URL + "/find_urn", json = params)
@@ -45,7 +52,8 @@ def ngram_book(
     ddk = None, 
     topic = None
 ):
-    """Get a time series for a word as string, title is name of book period is (year, year),
+    """Get a time series for a word as string, 
+    title is name of book period is (year, year),
     lang is three letter iso code.
     Use % as wildcard where appropriate - no wildcards in word and lang"""
     params = locals()
@@ -156,7 +164,7 @@ def document_corpus(
     if "ddk" in params:
         params["ddk"]  = "^" + params['ddk'].replace('.', '"."')
         
-    r=requests.post(BASE_URL + "/build_corpus", json=params)
+    r = requests.post(BASE_URL + "/build_corpus", json=params)
     
     return pd.DataFrame(r.json())
     
@@ -189,7 +197,7 @@ def totals(n = 50000):
 def concordance(urns = None, words = None, window = 25, limit = 100):
     """ Get a list of concordances from database, words is an fts5 string search expression"""
     if words is None:
-        return {}
+        return {} # exit condition
     else:
         params = {
             'urns': urns,
@@ -203,7 +211,7 @@ def concordance(urns = None, words = None, window = 25, limit = 100):
 def concordance_counts(urns = None, words = None, window = 25, limit = 100):
     """ Get a list of concordances from database, words is an fts5 string search expression"""
     if words is None:
-        return {}
+        return {} #exit condition
     else:
         params = {
             'urns': urns,
@@ -216,7 +224,7 @@ def concordance_counts(urns = None, words = None, window = 25, limit = 100):
 
 def konkordans(urns = None, query = None, window = 25, limit = 100):
     if query is None:
-        return {}
+        return {} # exit condition
     else:
         params = {
             'urns': urns,
